Umbau beim Proxy-Kram

+++ scripts/deployToProduction.sh
@@ -28,7 +28,11 @@ if (docker container ls -a | grep parcel-webservice-${deployment_color}-2); then
 fi
 docker run -d --restart always --network=parcelnetwork --name=parcel-webservice-${deployment_color}-1 ${asset_server}:443/parcel-api:0.$version
 docker run -d --restart always --network=parcelnetwork --name=parcel-webservice-${deployment_color}-2 ${asset_server}:443/parcel-api:0.$version
-docker stop parcel-proxy
-docker rm -f parcel-proxy
-docker run -d --restart always --network=parcelnetwork -p 8443:8443 -e STATE=$deployment_color --name=parcel-proxy ${asset_server}:443/parcel-proxy
+if (docker ps | grep parcel-proxy); then
+  docker stop parcel-proxy
+fi
+if (docker container ls -a | grep parcel-proxy); then
+  docker rm -f parcel-proxy
+fi
+docker run -d --restart always --network=parcelnetwork -p 8443:8443 -e STATE=${deployment_color} --name=parcel-proxy ${asset_server}:443/parcel-proxy
 eval $(docker-machine env -u)
